requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
huggingface-hub==0.20.3
datasets==2.18.0
python-dotenv==1.0.1
gunicorn==21.2.0
